+++ Sources/SuperwallKit/Analytics/Superwall Placement/SuperwallEventObjc.swift
@@ -131,6 +131,9 @@ public enum SuperwallEventObjc: Int, CaseIterable {
   /// When the request to load the paywall's products has failed and is being retried.
   case paywallProductsLoadRetry
 
+  /// When the paywall's products are missing from the App Store.
+  case paywallProductsLoadMissingProducts
+
   /// Information about a paywall presentation request
   case paywallPresentationRequest
 
@@ -289,6 +292,8 @@ public enum SuperwallEventObjc: Int, CaseIterable {
       return "paywallProductsLoad_fail"
     case .paywallProductsLoadRetry:
       return "paywallProductsLoad_retry"
+    case .paywallProductsLoadMissingProducts:
+      return "paywallProductsLoad_missingProducts"
     case .paywallProductsLoadComplete:
       return "paywallProductsLoad_complete"
     case .paywallPresentationRequest: